refactor: move FeeHistory to core types
